refactor(curator): rename lore_curator/core.py to curator_c.py
Part of Plan A / Task 4. The 2026-04-20 three-agent review surfaced
that lore_curator/core.py (~950 LOC of frontmatter hygiene, staleness
flagging, supersession detection) is Curator C in spirit but reads as
legacy overhang because the filename doesn't tell the triad story.

Pure rename. No code changes. __init__.py + __main__.py + three test
files have their imports redirected to the new name.

Baseline test count before: 750. After: 750. No behavioral change.

Grep acceptance:
  grep -rn 'from lore_curator import core\|lore_curator\.core' \
    lib/ tests/   → zero hits

Task 5 (next commit) will reframe the module docstring and optionally
rename run_curator → run_curator_c.
