+++ tests/test_open_items_migration.py
@@ -6,7 +6,7 @@ functions and covered separately by manual runs.
 
 from __future__ import annotations
 
-from lore_curator.core import (
+from lore_curator.curator_c import (
     extract_open_items,
     migrate_open_items,
 )
